Move story and test to ConnectDialog directory
